Move token helpers to utils
