+++ src/app/pages/RuntimeAccountDetailsPage/index.tsx
@@ -11,7 +11,7 @@ import { useRequiredScopeParam } from '../../hooks/useScopeParam'
 import { contractCodeContainerId } from './ContractCodeCard'
 import { useTokenInfo } from '../TokenDashboardPage/hook'
 import { accountTokenTransfersContainerId } from './AccountTokenTransfersCard'
-import { getTokenTypePluralName } from '../../../types/tokens'
+import { getTokenTypePluralName } from '../../utils/tokens'
 import { SearchScope } from '../../../types/searchScope'
 import { RuntimeAccountDetailsCard } from './RuntimeAccountDetailsCard'
 import { eventsContainerId } from './AccountEventsCard'
